tests/test_data_mixer.py
import numpy as np
import pytest
from src.data_mixer import DataMixer

def test_spatialise_signal_basic():
    """Verify spatialisation produces correct output shape and delay trend."""
    fs = 1000
    mixer = DataMixer(sample_rate=fs)
    
    # 1. Simple Linear Array along X
    mixer.mic_positions = np.array([
        [0.0, 0.0, 0.0],
        [1.0, 0.0, 0.0]  # 1 meter away
    ])
    mixer.speed_of_sound = 343.0
    
    # 2. Source from Azimuth 0 (X-axis)
    # Mic 2 is closer -> arrives earlier -> negative delay relative to origin (Mic 1)
    # Wait, origin is reference.
    # Mic 1 is at 0. Mic 2 is at 1.
    # Source is at infinity along +X.
    # Signal hits Mic 2 first. Then Mic 1.
    # Delay Mic 2 relative to Mic 1 should be negative (-1/c).
    
    fs = 1000
    signal = np.zeros(100)
    signal[50] = 1.0  # Impulse at t=50
    
    multi = mixer.spatialise_signal(signal, azimuth_deg=0.0)
    
    # Shape check
    assert multi.shape == (2, 100)
    
    # Peak check
    peak_mic1 = np.argmax(np.abs(multi[0]))
    peak_mic2 = np.argmax(np.abs(multi[1]))
    
    # Mic 2 is closer, should arrive earlier (smaller index)
    assert peak_mic2 < peak_mic1
    
    # Expected sample diff
    # d = 1m. t = 1/343 = 0.0029s. Samples = 0.0029 * 1000 = 2.9 samples.
    # Since we use sub-sample shift, peak might be at 47 or 48 vs 50.
    diff = peak_mic1 - peak_mic2
    assert 2 <= diff <= 4

def test_mix_signals_snr():
    """Verify mixing maintains SNR logic."""
    mixer = DataMixer()
    
    sig1 = np.random.randn(1000)
    sig2 = np.random.randn(1000)
    
    # Mix with 0dB SNR -> Power should be equal
    mixed = mixer.mix_signals(sig1, sig2, snr_db=0.0)
    
    # Check if mixed signal is roughly combination
    # Hard to test exactly without strict scaling, but let's check non-zero
    assert np.max(np.abs(mixed)) > 0
    assert mixed.shape == sig1.shape
