requirements.txt
click
pydantic
PyYAML
librosa
torch
scipy
numpy
soundfile
matplotlib
